fix: use countdown hook
