+++ src/types/timeLeft.ts
@@ -1,0 +1,6 @@
+export interface TimeLeft {
+	days: number;
+	hours: number;
+	minutes: number;
+	seconds: number;
+}
